+++ backend/budget/views.py
@@ -1,3 +1,18 @@
-from django.shortcuts import render
+from rest_framework import viewsets
+from rest_framework.response import Response
+from rest_framework.permissions import IsAuthenticated
+from rest_framework.decorators import action
 
-# Create your views here.
+
+class BudgetViewSet(viewsets.ViewSet):
+    permission_classes = (IsAuthenticated,)
+    
+    @action(
+        detail=False,
+        methods=['get'],
+        permission_classes=[IsAuthenticated],
+        url_path='budgets',
+    )
+    def get_budgets(self, request):
+        content = {'message': 'Hello, World!'}
+        return Response(content)