+++ frontend/src/api/api.tsx
@@ -2,6 +2,7 @@ import axios from 'axios';
 import cookie from 'cookie';
 
 const api = axios.create();
+
 api.interceptors.request.use((config) => {
   const { csrftoken } = cookie.parse(document.cookie);
   if (csrftoken) {