+++ frontend/src/pages/landing/components/Login.tsx
@@ -1,12 +1,14 @@
 import React from "react";
 import { Button, Form } from "react-bootstrap";
 import { useForm } from "react-hook-form";
-import { login, LoginCredentials } from "../../../api/login";
+import { creators } from "../../../api/auth/login";
 import Email from "./components/Email";
 import FormSwitch from "./components/FormSwitch";
 import Password from "./components/Password";
 import { RootStateOrAny, useDispatch, useSelector } from "react-redux";
 import { Navigate } from "react-router-dom";
+import { CredentialsData, LoginCredentials } from "../../../types/user";
+
 
 interface Props {
   setLoginForm: (loginForm: boolean) => void;
@@ -19,16 +21,18 @@ const Login = ({ setLoginForm }: Props) => {
     formState: { errors },
   } = useForm<LoginCredentials>();
 
-  const restLogin = useSelector((state: RootStateOrAny) => state.restLogin);
+  const loginData: CredentialsData  = useSelector((state: RootStateOrAny) => state.login);
   const dispatch = useDispatch();
 
   const onSubmit = (data: LoginCredentials) => {
     // Sending password in plaintext (even when SSL is used) is a questionable practice.
     // https://cheatsheetseries.owasp.org/cheatsheets/REST_Security_Cheat_Sheet.html
-    dispatch(login.post(data));
+    dispatch(creators.login(data));
   };
 
-  if (restLogin.verified) {
+  console.log(loginData);
+
+  if (loginData.token) {
     return <Navigate replace to="/budgets" />;
   }