+++ frontend/src/store/reducers.tsx
@@ -1,17 +1,17 @@
 import { connectRouter } from 'connected-react-router';
 import { BrowserHistory } from 'history';
 import { combineReducers } from 'redux';
-
 import { restCheckReducer as restCheck } from '../api/rest_check';
-import { restSendExpense as expense } from '../api/expense';
-import { restLogin } from '../api/login';
+import { expenseReducer as expense } from '../api/expense';
+import { loginReducer } from '../api/auth/login';
+import { registerReducer } from '../api/auth/register';
 
 export const createRootReducer = (history: BrowserHistory) => {
   return combineReducers({
     router: connectRouter(history),
-    restLogin,
-    restCheck,
-    expense,
-
+    login: loginReducer,
+    register: registerReducer,
+    restCheck: restCheck,
+    expense: expense,
   });
 };