+++ docs/CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [[0.4.0-alpha]] - 2022-03-24
+
+###### _([diff: 0.3.5-0.4.0-alpha])_
+
+### Added
+
+- Token Authorization System
+- Persistency via storing `token` in `localStorage`
+- Login ([back](../backend/users/views/login.py)/[front](../frontend/src/pages/landing/components/Login.tsx)) and Register ([back](../backend/users/views/login.py)/[front](../frontend/src/pages/landing/components/Register.tsx)) - API communication
+- Password encryption via salted **unsafe** (!) [sha256](https://pl.wikipedia.org/wiki/SHA-2) algorithm on both frontend and backend
+
 ## [[0.3.5]] - 2022-03-24
 
 ###### _([diff: 0.3.4-0.3.5])_
@@ -179,6 +190,7 @@ Connected Frontend & Backend via webpack_loader.
     -   I know I should implement them right away and have all of that sort of things already done, but I _really_ want to start some dev coding 🐈
     -   ... still got to setup frontend-backend, so I'll prioritize that at this moment.
 
+[diff: 0.3.5-0.4.0-alpha]: https://github.com/Luzkan/FamilyBudget/compare/0.3.5...0.4.0-alpha
 [diff: 0.3.4-0.3.5]: https://github.com/Luzkan/FamilyBudget/compare/0.3.4...0.3.5
 [diff: 0.3.3-0.3.4]: https://github.com/Luzkan/FamilyBudget/compare/0.3.3...0.3.4
 [diff: 0.3.2-0.3.3]: https://github.com/Luzkan/FamilyBudget/compare/0.3.2...0.3.3
@@ -190,6 +202,7 @@ Connected Frontend & Backend via webpack_loader.
 [diff: 0.1.1-0.2.0]: https://github.com/Luzkan/FamilyBudget/compare/0.1.1...0.2.0
 [diff: 0.1.0-0.1.1]: https://github.com/Luzkan/FamilyBudget/compare/0.1.0...0.1.1
 [diff: 0.0.0-0.1.0]: https://github.com/Luzkan/FamilyBudget/compare/0.0.0...0.1.0
+[0.4.0-alpha]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.4.0-alpha
 [0.3.5]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.3.5
 [0.3.4]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.3.4
 [0.3.3]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.3.3